requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
msgpack==1.1.1
packaging==25.0
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
